requirements.txt
# Core Framework
Flask==3.0.2
Flask-SocketIO==5.3.6
Flask-CORS==4.0.1
Flask-JWT-Extended==4.6.0

# Database
pymongo==4.7.3
redis==5.0.7

# Cryptography & Security
cryptography==45.0.0
bcrypt==4.1.3
Argon2-cffi==23.1.0

# Face Recognition
face-recognition==1.3.0
opencv-python==4.10.0.84
Pillow==10.4.0

# Authentication
PyJWT==2.8.0

# File Handling
python-magic==0.4.27
Werkzeug==3.0.3

# Environment
python-dotenv==1.0.1

# Production
gunicorn==22.0.0
